# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/v1/inference/project/{owner}/{pid}": {
            "get": {
                "description": "create inference",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Inference"
                ],
                "summary": "Create",
                "parameters": [
                    {
                        "type": "string",
                        "description": "project owner",
                        "name": "owner",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "project id",
                        "name": "pid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/app.InferenceDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "bad_request_body"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "bad_request_param"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "system_error"
                        }
                    }
                }
            }
        },
...
        "/v1/user/following": {
            "post": {
                "description": "add a following",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Following"
                ],
                "parameters": [
                    {
                        "description": "body of creating following",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.followingCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "bad_request_body"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "bad_request_param"
                        }
                    },
                    "402": {
                        "description": "Payment Required",
                        "schema": {
                            "type": "not_allowed"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "resource_not_exists"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "duplicate_creating"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "system_error"
                        }
                    }
                }
            }
        },
        "/v1/user/following/{account}": {
            "get": {
                "description": "list followings",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Following"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "the account the followings belong to",
                        "name": "account",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.FollowsDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "system_error"
                        }
                    }
                }
            },
            "delete": {
                "description": "remove a following",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Following"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "the account of following",
                        "name": "account",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "bad_request_param"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "not_allowed"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "system_error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "app.FollowDTO": {
            "type": "object",
            "properties": {
                "account": {
                    "type": "string"
                },
                "avatar_id": {
                    "type": "string"
                },
                "bio": {
                    "type": "string"
                },
                "is_follower": {
                    "type": "boolean"
                }
            }
        },
        "app.FollowsDTO": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/app.FollowDTO"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "app.InferenceDTO": {
            "type": "object",
            "properties": {
                "access_url": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "inference_id": {
                    "type": "string"
                }
            }
        },
...
        "controller.followingCreateRequest": {
            "type": "object",
            "properties": {
                "account": {
                    "type": "string"
                }
            }
        },
...
    }
}